+++ components/Case/types.d.ts
@@ -1,0 +1,12 @@
+export type Tech =
+  | 'next'
+  | 'typescript'
+  | 'css'
+  | 'react'
+  | 'html'
+  | 'motion'
+  | 'tailwind'
+  | 'github'
+  | 'git'
+  | 'node'
+  | 'supabase';
